% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{gen.child.list.mst}
\alias{gen.child.list.mst}
\title{gen.child.list.mst}
\usage{
gen.child.list.mst(clust.edge.list, m)
}
\arguments{
\item{clust.edge.list}{The return value of the kruskal() function.}

\item{m}{Number of nodes.}
}
\value{
An adjacency list in the form of a list of vectors.
}
\description{
Generate an adjacency list
}
