% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{is.same.set}
\alias{is.same.set}
\title{is.same.set}
\usage{
is.same.set(i, j, ufds)
}
\arguments{
\item{i}{The first element in the tuple.}

\item{j}{The second element in the tuple.}

\item{ufds}{A data.table representing a UFDS.}
}
\value{
TRUE if the elements are in the same set, FALSE otherwise.
}
\description{
Check if two elements are in the same set
}
