% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{kruskal}
\alias{kruskal}
\title{kruskal}
\usage{
kruskal(edge.list, m)
}
\arguments{
\item{edge.list}{A data frame with columnns 'from', 'to', 'dist'.}

\item{m}{Number of nodes.}
}
\value{
A list of edges in the MST, in the same format as the input argument edge.list.
}
\description{
Kruskal's algorithm for MST computation.
}
