% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{mst.cluster}
\alias{mst.cluster}
\title{Run clustering using MST.
Before calling this function, remove some edges from the MST, for example the  k-1 heaviest.
For example:\preformatted{k <- 3
n.edges <- nrow(mst.edge.list)
child.list.mst <- child.list.mst[1:(n.edges - (k - 1)),]
}}
\usage{
mst.cluster(child.list.mst, m, k)
}
\arguments{
\item{child.list.mst}{The return value of the gen.child.list.mst() function with k-1 edges removed.}

\item{m}{Number of nodes.}

\item{k}{The number of clusters.}
}
\value{
A vector whose k-th element is the cluster the k-th point belongs to.
}
\description{
Run clustering using MST.
Before calling this function, remove some edges from the MST, for example the  k-1 heaviest.
For example:\preformatted{k <- 3
n.edges <- nrow(mst.edge.list)
child.list.mst <- child.list.mst[1:(n.edges - (k - 1)),]
}
}
