% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{reset.ufds}
\alias{reset.ufds}
\title{Initialize UFDS}
\usage{
reset.ufds(m)
}
\arguments{
\item{m}{Number of elements.}
}
\value{
An instance of the class ufds containing m disjoint sets.
}
\description{
Initialize UFDS
}
