% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\docType{class}
\name{ufds-class}
\alias{ufds-class}
\title{UFDS}
\arguments{
\item{p}{parent of an item, initially the item itself}

\item{rank}{upper bound to tree depth excluding root (initially 0)}
}
\description{
UFDS
}

