% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{union.set}
\alias{union.set}
\title{union.set}
\usage{
union.set(i, j, ufds)
}
\arguments{
\item{i}{The first element in the tuple.}

\item{j}{The second element in the tuple.}

\item{ufds}{A data.table representing a UFDS.}
}
\value{
No return value, called for side effects on rank and p.
}
\description{
Join the sets the two elements passed as arguments belong to.
}
