% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{find.set}
\alias{find.set}
\title{Find the set an elements belongs to.}
\usage{
find.set(i, ufds)
}
\arguments{
\item{i}{The element to check.}

\item{ufds}{An instance of the class ufds.}
}
\description{
Find the set an elements belongs to.
}
