% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{find.set}
\alias{find.set}
\title{find.set}
\usage{
find.set(i, ufds)
}
\arguments{
\item{i}{The element to check.}

\item{ufds}{A data.table representing a UFDS.}
}
\value{
An integer: the root node of the set the element belongs to.
}
\description{
Find the set an element belongs to.
}
