% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstclustering.R
\name{reset.ufds}
\alias{reset.ufds}
\title{reset.ufds}
\usage{
reset.ufds(m)
}
\arguments{
\item{m}{Number of elements.}
}
\value{
A data table containing a 'rank' column and a 'parent' column.
}
\description{
Initialize UFDS
}
